Fix wrong log move codec implementation to go-micro
