+++ client/highway/highway_client.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	"github.com/ServiceComb/go-chassis/core/client"
-	"github.com/ServiceComb/go-chassis/core/codec"
+	"github.com/ServiceComb/go-chassis/third_party/forked/go-micro/codec"
 	"github.com/ServiceComb/go-chassis/core/config"
 	"github.com/ServiceComb/go-chassis/core/lager"
 	microClient "github.com/ServiceComb/go-chassis/third_party/forked/go-micro/client"
